Implemented JPA entities for portfolio management system
