+++ src/main/java/com/wellsfargo/counselor/entity/Portfolio.java
@@ -1,0 +1,47 @@
+package com.wellsfargo.counselor.entity;
+
+
+import jakarta.persistence.*;
+
+@Entity
+public class Portfolio {
+
+    @Id
+    @GeneratedValue()
+    private long portfolioId;
+
+    @ManyToOne
+    private Client client;
+
+    @Column(nullable = false)
+    private String creationDate;
+
+    protected Portfolio() {
+
+    }
+
+    public Portfolio(Client client, String creationDate) {
+        this.client = client;
+        this.creationDate = creationDate;
+    }
+
+    public Long getPortfolioId() {
+        return portfolioId;
+    }
+
+    public Client getCLient() {
+        return client;
+    }
+
+    public void setClient(Client client) {
+        this.client = client;
+    }
+
+    public String getCreationDate() {
+        return creationDate;
+    }
+
+    public void setCreationDate(String creationDate) {
+        this.creationDate = creationDate;
+    }
+}
